Refactored backend evaluation benchmarks
